Test for internal TLS certificate rotation (#15217) (#732)
* TestTLSCertificateRotation

* Use SkipCleanupOnFail only when the test fails

The original intention was to skip the cleanup only when the test fail,
not always. This was missed in the original PR that introduced the flag.

* Use default poll interval and timeout

* Fix lint

* Make function GetCASecret return the secret also for internal encryption

* Move WaitForLog to e2e package to prevent "redefined flag"

* Wait for 2 occurrences of cert reload

It is loaded at the beginning and then after re-creating the
serving-certs secret during the test.

* Increase timeout for systeminternaltls tests to 4m

* Wait for CA cert to be re-created in routing-serving-certs before checking state

* The test usually takes around 120 seconds

* Rename waitForCaCert to waitForCACertSecret

* Fix lint

* Use new ClusterIssuer for testing certificate rotation

Also, do not rely on ca.crt to be in routing-serving-certs. We rely
solely on the trust-bundle ConfigMap.

* Actually rename the new issuer

* Run ./hack/update-codegen.sh

* Create const for name of renewed issuer

* Fixes

* Update test

* Revert "Use new ClusterIssuer for testing certificate rotation"

* Fix lint and deps

* Use arrays for GO_TEST_FLAGS and E2E_TEST_FLAGS

This is to properly expand the variables when passing to tests and to
pass shellcheck.

* One more fix for shellcheck

* Add missing bracket

* Do NOT delete old cert from trust-bundle ConfigMap

The cert rotation might not be fully complete yet. This also adds a
little randomness to the test.

(cherry picked from commit b15ce9a57a86f2bb73a7c35fe01d17d0cc89da1e)

Co-authored-by: Martin Gencur <mgencur@redhat.com>