Do not run TestTLSCertificateRotation in parallel with other tests (#776)
This test deletes the serving-tests/serving-certs during the test run.
This might break the other test TestSystemInternalTLS because it waits
for certain log lines to appear in the service pods. Deleting the
serving-certs secret causes the application pods to be deleted so that
the new cert can be mounted. And waiting for given log lines then fails
because the pod no longer exists:
system_internal_tls_test.go:110: TLS not used on requests to
queue-proxy: pods
"system-internal-tls-wgvxqtyx-00001-deployment-6fdc4c8c7d-4ll8q" not
found